Refresh Daemon
Simplify by relying on jars to effect the actual refresh. The login in refresh.php has been ported to jars itself.

+++ jars-refreshd.sh
@@ -5,8 +5,8 @@ dir="$(dirname "$0")"
 source "$dir/functions.sh"
 
 if [ "refresh/conf.d/*.conf" == "refresh/conf.d/\\*.conf" ]; then
-    echo "No configs to monitor"
-    exit
+    >&2 echo "No configs to monitor"
+    exit 1
 fi
 
 for conf_file in refresh/conf.d/*.conf; do
@@ -22,11 +22,13 @@ for key in ${!WATCH_FILES[@]}; do
 done
 
 for key in ${!CONNECTION_STRINGS[@]}; do
-    BIN_HOME=${BIN_HOMES[$key]}
-    refresh_portal ${NAMES[$key]}
-done
+    export AUTH_TOKEN="${AUTH_TOKENS[$key]}"
+    export BIN_HOME="${BIN_HOMES[$key]}"
+    export CONNECTION_STRING="${CONNECTION_STRINGS[$key]}"
+    export PORTAL_AUTOLOAD="${PORTAL_AUTOLOADS[$key]}"
 
-BIN_HOME=
+    refresh_portal
+done
 
 rm -rf /tmp/jars-refreshd.fifo
 mkfifo /tmp/jars-refreshd.fifo
@@ -41,15 +43,24 @@ while true; do
     PID=$!
 
     while read f e; do
+        FOUND=
+
         for key in ${!WATCH_FILES[@]}; do
             if [ "${WATCH_FILES[$key]}" == "$f" ]; then
-                AUTH_TOKEN="${AUTH_TOKENS[$key]}"
-                BIN_HOME="${BIN_HOMES[$key]}"
-                CONNECTION_STRING="${CONNECTION_STRINGS[$key]}"
-                NAME="${NAMES[$key]}"
-                WATCH_FILE="${WATCH_FILES[$key]}"
+                FOUND=1
+
+                export AUTH_TOKEN="${AUTH_TOKENS[$key]}"
+                export BIN_HOME="${BIN_HOMES[$key]}"
+                export CONNECTION_STRING="${CONNECTION_STRINGS[$key]}"
+                export PORTAL_AUTOLOAD="${PORTAL_AUTOLOADS[$key]}"
+                export WATCH_FILE="${WATCH_FILES[$key]}"
             fi
         done
+
+        if [ -z "$FOUND" ]; then
+            >&2 echo "Internal error"
+            exit 1
+        fi
 
         if [[ "$OSTYPE" =~ ^darwin ]] && [[ "$e" =~ Renamed ]] || [ "$e" == "DELETE_SELF" ] ; then
             kill $PID
@@ -58,10 +69,10 @@ while true; do
                 sleep 1
             done
 
-            refresh_portal $NAME
+            refresh_portal
             break
         fi
 
-        refresh_portal $NAME
+        refresh_portal
     done < /tmp/jars-refreshd.fifo
 done
